Change element render approach
Instead of complicated approaches to merge newlines between a trait /
special title and body,
we now prepend the name to the body using markdown italic tags and let
the markdown engine figure it out.

All we then have to do is add minor,
clear CSS to indent everything but the very first line.